+++ ESP32/src/LoadCell.cpp
@@ -55,12 +55,12 @@ ADS1256& ADC() {
 
 void LoadCell_ADS1256::setLoadcellRating(uint8_t loadcellRating_u8) const {
   ADS1256& adc = ADC();
-  double originalConversionFactor_f64 = CONVERSION_FACTOR;
+  float originalConversionFactor_f64 = CONVERSION_FACTOR;
   
-  double updatedConversionFactor_f64 = 1;
+  float updatedConversionFactor_f64 = 1;
   if (LOADCELL_WEIGHT_RATING_KG>0)
   {
-      updatedConversionFactor_f64 = 2 * ((double)loadcellRating_u8) * (CONVERSION_FACTOR/LOADCELL_WEIGHT_RATING_KG);
+      updatedConversionFactor_f64 = 2 * ((float)loadcellRating_u8) * (CONVERSION_FACTOR/LOADCELL_WEIGHT_RATING_KG);
   }
   Serial.print("OrigConversionFactor: ");
   Serial.print(originalConversionFactor_f64);
